# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Comoon API
  description: Social commerce platform connecting leaders, entrepreneurs, and causes in Colombia.
  version: 1.0.0
  contact:
    name: Comoon Team
servers:
  - url: https://comoon.pages.dev/api
    description: Production
  - url: http://localhost:4321/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    UserSession:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, leader, entrepreneur]
        profileId:
          type: integer
        name:
          type: string
        isVerified:
          type: boolean
        isActive:
          type: boolean
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserSession"
        redirectTo:
          type: string
        error:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, confirmPassword, role, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        confirmPassword:
          type: string
        role:
          type: string
          enum: [leader, entrepreneur]
        name:
          type: string
        storeName:
          type: string
          description: Required when role is entrepreneur
        location:
          type: string
        bio:
          type: string
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        userId:
          type: integer
        error:
          type: string
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
  /auth/login:
    post:
      summary: Log in with email and password
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful (sets HttpOnly cookie)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /auth/me:
    get:
      summary: Get current authenticated user profile
      tags: [Auth]
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  email:
                    type: string
                  role:
                    type: string
                  profileId:
                    type: integer
                  name:
                    type: string
                  photoUrl:
                    type: string
                  bio:
                    type: string
                  location:
                    type: string
                  isVerified:
                    type: boolean
                  isActive:
                    type: boolean
                  dashboardPath:
                    type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserSession"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/change-password:
    post:
      summary: Change password for authenticated user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword, confirmPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                confirmPassword:
                  type: string
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...